backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
torch
torchvision
transformers
timm
pillow
numpy
scikit-learn
python-dotenv
pinecone
pinecone[asyncio]
modal
boto3
botocore
open-clip-torch
python-jose[cryptography]
passlib[bcrypt]
motor
pymongo
pydantic
pydantic-settings
email-validator
beanie